refactor: use next.js link and move assets to public (#15)
* fix: use next link and move images to public folder

* refactor: utilize alias for root directory

* refactor: use alias for discover page
